pyproject.toml
[tool.poetry]
name = "jun-hackathon"
version = "0.1.0"
description = ""
authors = ["VadimK <yaralavadim@gmail.com>"]
readme = "README.md"
packages = [{include = "jun_hackathon"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.12"
psycopg2 = "^2.9.9"
pydantic = "^2.4.2"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
